requirements.txt
# Core dependencies
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
pymysql>=1.1.0
pandas>=2.0.0
numpy>=1.24.0

# LLM Integration
openai>=1.12.0
tiktoken>=0.5.0

# Data Analysis
scipy>=1.11.0

# CLI and Configuration
click>=8.1.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
rich>=13.0.0

# Output Generation
jinja2>=3.1.0
markdown>=3.5.0

# Progress and Logging
tqdm>=4.66.0
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Type hints
typing-extensions>=4.8.0
